* Add new activity for interacting with Persons. * Add ViewModel to invoke Interactors. * Add my own MockUp repository with generation of Persons. * Create functions to get persons list and change the state of persons using LiveData.

+++ app/src/main/java/com/kozak/pw/domain/news/ReadNewsUseCase.kt
@@ -3,5 +3,5 @@ package com.kozak.pw.domain.news
 import com.kozak.pw.domain.repository.NewsItemRepository
 
 class ReadNewsUseCase(private val newsItemRepository: NewsItemRepository) {
-    fun readNews(newsItemId: Int) = newsItemRepository.readNews(newsItemId)
+    fun readNews(newsItemId: Long) = newsItemRepository.readNews(newsItemId)
 }
